zenohc examples enabled
