+++ CMakeLists.txt
@@ -66,5 +66,5 @@ endif()
 if(${CMAKE_SOURCE_DIR} STREQUAL ${CMAKE_CURRENT_SOURCE_DIR})
 	# add_subdirectory(install)
 	add_subdirectory(tests)
-	# add_subdirectory(examples)
+	add_subdirectory(examples)
 endif()
