+++ examples/CMakeLists.txt
@@ -40,11 +40,13 @@ function(add_examples glob mode lib is_cpp)
 endfunction()
 
 # C examples compiled with C compiler and zenohc
-add_examples("${CMAKE_CURRENT_SOURCE_DIR}/c/*.c" "c" zenohc::lib FALSE)
+add_examples("${CMAKE_CURRENT_SOURCE_DIR}/c/*.c" c zenohc::lib FALSE)
 # C examples compiled with C++ compiler and zenohc
-add_examples("${CMAKE_CURRENT_SOURCE_DIR}/c/*.c" "cpp_c" zenohc::lib FALSE)
+add_examples("${CMAKE_CURRENT_SOURCE_DIR}/c/*.c" cpp_c zenohc::lib FALSE)
 
-# C++ examples compiled with C++ compiler with zenohc library
-add_examples("${CMAKE_CURRENT_SOURCE_DIR}/cpp/*.cpp" cpp_zenohc zenohcpp::zenohc::lib TRUE)
-# C++ examples compiled with C++ compiler with zenohpico library
-add_examples("${CMAKE_CURRENT_SOURCE_DIR}/cpp/*.cpp" cpp_zenohpico zenohcpp::zenohpico TRUE)
+# C++ examples compiled with C++ compiler with zenohc library only
+add_examples("${CMAKE_CURRENT_SOURCE_DIR}/cpp/zenohc/*.cpp" cpp_zenohc zenohcxx::zenohc::lib TRUE)
+
+# C++ examples compiled with C++ compiler with both zenohc and zenohpico libraries
+add_examples("${CMAKE_CURRENT_SOURCE_DIR}/cpp/universal/*.cpp" cpp_zenohc zenohcxx::zenohc::lib TRUE)
+# add_examples("${CMAKE_CURRENT_SOURCE_DIR}/cpp/universal/*.cpp" cpp_zenohpico zenohcxx::zenohpico TRUE)
