+++ examples/cpp/universal/z_delete.cxx
@@ -14,9 +14,13 @@
 #include <iostream>
 
 #include "stdio.h"
-#include "zenohcpp.h"
+#include "zenoh.hxx"
 
-using namespace zenoh;
+#if defined(ZENOHCXX_ZENOHPICO)
+using namespace zenohpico;
+#elif defined(ZENOHCXX_ZENOHC)
+using namespace zenohc;
+#endif
 
 int _main(int argc, char **argv) {
     /*    const char *keyexpr = "demo/example/zenoh-cpp-put";
