+++ examples/cpp/universal/z_get.cxx
@@ -18,9 +18,13 @@
 #include <iostream>
 #include <mutex>
 
-#include "zenohcpp.h"
+#include "zenoh.hxx"
 
-using namespace zenoh;
+#if defined(ZENOHCXX_ZENOHPICO)
+using namespace zenohpico;
+#elif defined(ZENOHCXX_ZENOHC)
+using namespace zenohc;
+#endif
 
 int _main(int argc, char **argv) {
     const char *expr = "demo/example/**";
