+++ examples/cpp/universal/z_scout.cxx
@@ -24,9 +24,13 @@
 #include <unistd.h>
 #endif
 
-#include "zenohcpp.h"
+#include "zenoh.hxx"
 
-using namespace zenoh;
+#if defined(ZENOHCXX_ZENOHPICO)
+using namespace zenohpico;
+#elif defined(ZENOHCXX_ZENOHC)
+using namespace zenohc;
+#endif
 
 void printlocators(const StrArrayView &locs) {
     std::cout << "[";
