+++ examples/cpp/universal/z_sub_thr.cxx
@@ -14,9 +14,13 @@
 #include <stdio.h>
 #include <time.h>
 
-#include "zenohcpp.h"
+#include "zenoh.hxx"
 
-using namespace zenoh;
+#if defined(ZENOHCXX_ZENOHPICO)
+using namespace zenohpico;
+#elif defined(ZENOHCXX_ZENOHC)
+using namespace zenohc;
+#endif
 
 #define N 1000000
